refactor(apps): extract DenoSubprocessController logic to BaseSubprocessController
